fix: address remaining compiler review cleanup
