+++ apps/frontend-demo/src/components/molecules/header-search.tsx
@@ -82,9 +82,7 @@ export function HeaderSearch() {
       if (nextSelectedValue === "__search__") {
         handleSearch(searchQuery)
       } else if (
-        searchResults.some(
-          (product) => (product.handle || product.id) === nextSelectedValue
-        )
+        searchResults.some((product) => product.handle === nextSelectedValue)
       ) {
         router.push(`/products/${nextSelectedValue}`)
       } else {
