+++ apps/herbatika/src/components/product-detail.tsx
@@ -92,7 +92,7 @@ export function ProductDetail({ handle }: ProductDetailProps) {
       controller.productQuery.isLoading ||
       controller.productQuery.error ||
       controller.product ? null : (
-        <section className="space-y-400 rounded-sm border border-border-secondary bg-surface p-600">
+        <section className="space-y-400 rounded-base border border-border-secondary bg-surface p-600">
           <StatusText showIcon status="error">
             Produkt sa nepodarilo nájsť.
           </StatusText>
